// pokemon-app/src/app/models/pokemonProfile.ts
export class PokemonProfile {
  height: number;
  weight: number;
  catchRate: number;
  genderRatio: string;
  eggGroups: string;
  hatchSteps: string;
  abilities: string;
  EVs: string;
  color: string;
  id: number;
  evolution: Array<string>;
  name: string;
  mainPokemonAvatar: string;
  evolvedPokemonAvatar: string;
  evolutionChain: string;
  hp: number;
  defense: number;
  specialAttack: number;
  specialDefense: number;
  speed: number;
  attack: any;
  types: Array<string>;
}